snowflake_project/requirements.txt
snowflake
python-dotenv
